# open-api/kondis-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/upload/activity": {
      "post": {
        "operationId": "UploadController_uploadActivity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": ".fit, .tcx, or .gpx activity file"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Activity file processing is queued and happens asynchronously",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitUploadResponseDto_Output"
                }
              }
            }
          }
        },
        "summary": "Upload a FIT, TCX, or GPX activity file",
        "tags": [
          "uploads"
        ]
      }
    },
    "/upload/strava": {
      "post": {
        "operationId": "UploadController_uploadStravaTakeout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Strava takeout .zip file"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Takeout importing and activity parsing are queued and happen asynchronously",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LagomTakeoutUploadResponseDto_Output"
                }
              }
            }
          }
        },
        "summary": "Import activities from a Strava takeout ZIP archive",
        "tags": [
          "uploads"
        ]
      }
    },
    "/upload/strava/{id}": {
      "get": {
        "operationId": "UploadController_getStravaTakeoutStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current import progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TakeoutImportStatusDto_Output"
                }
              }
            }
          }
        },
        "summary": "Get Strava takeout import progress",
        "tags": [
          "uploads"
        ]
      }
    },
...
    "/activities/{id}/images": {
      "post": {
        "operationId": "ActivityImageController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "caption": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image processing was queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityImageDto_Output"
                }
              }
            }
          }
        },
        "summary": "Upload an image to an activity",
        "tags": [
          "activity-images"
        ]
      },
      "get": {
        "operationId": "ActivityImageController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity images",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityImageListDto_Output"
                }
              }
            }
          }
        },
        "summary": "List ready images attached to an activity",
        "tags": [
          "activity-images"
        ]
      }
    },
    "/activities/{activityId}/images/{imageId}": {
      "patch": {
        "operationId": "ActivityImageController_update",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityImageUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated activity image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityImageDto_Output"
                }
              }
            }
          }
        },
        "tags": [
          "activity-images"
        ]
      },
      "delete": {
        "operationId": "ActivityImageController_delete",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "activity-images"
        ]
      }
    },
    "/activity-images/{imageId}/{variant}": {
      "get": {
        "operationId": "ActivityImageController_file",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variant",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read an image variant",
        "tags": [
          "activity-images"
        ]
      }
    },
...
    "/users/{id}/avatar": {
      "get": {
        "operationId": "UserController_avatarFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kondis API",
    "description": "OpenAPI schema for the Kondis server",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "schemas": {
...
      "FitUploadResponseDto_Output": {
        "type": "object",
        "properties": {
          "byteSize": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Uploaded activity file size in bytes"
          },
          "queued": {
            "type": "boolean",
            "description": "True when activity processing was submitted to the queue",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "byteSize",
          "queued"
        ],
        "additionalProperties": false
      },
      "LagomTakeoutUploadResponseDto_Output": {
        "type": "object",
        "properties": {
          "byteSize": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Uploaded takeout size in bytes"
          },
          "queued": {
            "type": "boolean",
            "description": "True when the takeout import was submitted to the queue",
            "enum": [
              true
            ]
          },
          "importId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "Identifier used to poll import progress"
          }
        },
        "required": [
          "byteSize",
          "queued",
          "importId"
        ],
        "additionalProperties": false
      },
      "TakeoutImportStatusDto_Output": {
        "type": "object",
        "properties": {
          "importId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "completed",
              "failed"
            ]
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "nullable": true
          },
          "processed": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "failed": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "duplicates": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "importId",
          "status",
          "total",
          "processed",
          "failed",
          "duplicates",
          "error"
        ],
        "additionalProperties": false
      },
...
      "ActivityImageDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "caption": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "nullable": true
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "ready",
              "failed"
            ]
          },
          "thumbnail": {
            "type": "string",
            "nullable": true
          },
          "preview": {
            "type": "string",
            "nullable": true
          },
          "original": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "caption",
          "sortOrder",
          "width",
          "height",
          "status",
          "thumbnail",
          "preview",
          "original"
        ],
        "additionalProperties": false
      },
...
      "ActivityImageListDto_Output": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ActivityImageDto_Output"
        }
      },
      "ActivityImageUpdateDto": {
        "type": "object",
        "properties": {
          "caption": {
            "type": "string",
            "maxLength": 10000,
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        }
      },
...
    }
  }
}